Update OpenMV.py
This is the latest script made for the OpenMV camera. It resets the "I' and "emailNum" values to zero when the Raspberry Pi script sends "snap." This then will have the OpenMV camera send a value of 244 (because I = 0) to the Raspberry Pi, which then puts the system back into panning motion. The EmailNum value being set to zero will allow the OpenMV camera to send image data for an email to be sent. Once an email is sent, EmailNum will equal one, which will not change unless "snap" is sent from the Raspberry Pi.